+++ NEWS.md
@@ -1,3 +1,8 @@
+# PEF 3.2.1
+
+- Remove computing sds outside the loops when operational search is used.
+- Add demo for using operational search.
+
 # PEF 3.2.0
 
 - Synchronize versions of all components.